# specs/openapi.baseline.json
{
  "components": {
...
    "responses": {
...
      "UnsupportedMediaType": {
        "description": "Unsupported Media Type"
      }
    },
    "schemas": {
...
      "ImportErrorResponse": {
        "properties": {
          "error": {
            "enum": [
              "invalid",
              "unsupported",
              "rate_limited"
            ],
            "type": "string"
          },
          "limit": {
            "nullable": true,
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "ok": {
            "enum": [
              false
            ],
            "type": "boolean"
          },
          "retry_after": {
            "nullable": true,
            "type": "integer"
          }
        },
        "required": [
          "ok",
          "error",
          "message"
        ],
        "type": "object"
      },
      "ImportMenuRequest": {
        "properties": {
          "items": {
            "items": {
              "properties": {
                "name": {
                  "type": "string"
                }
              },
              "required": [
                "name"
              ],
              "type": "object"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "ImportOkResponse": {
        "additionalProperties": false,
        "properties": {
          "diff": {
            "description": "Menu dry-run diff entries (menu only)",
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "dry_run": {
            "description": "Legacy alias (deprecated); prefer meta.dry_run",
            "type": "boolean"
          },
          "meta": {
            "additionalProperties": false,
            "properties": {
              "count": {
                "type": "integer"
              },
              "dry_run": {
                "description": "Present and true when request used ?dry_run=1",
                "type": "boolean"
              },
              "format": {
                "description": "Detected import format",
                "enum": [
                  "csv",
                  "docx",
                  "xlsx",
                  "menu"
                ],
                "type": "string"
              }
            },
            "required": [
              "count"
            ],
            "type": "object"
          },
          "ok": {
            "enum": [
              true
            ],
            "type": "boolean"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ImportRow"
            },
            "type": "array"
          }
        },
        "required": [
          "ok",
          "rows",
          "meta"
        ],
        "type": "object"
      },
      "ImportRow": {
        "properties": {
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "priority"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Unified Platform API",
    "version": "0.3.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/import/csv": {
      "post": {
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportOkResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportErrorResponse"
                }
              }
            },
            "description": "Invalid"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportErrorResponse"
                }
              }
            },
            "description": "Rate limited"
          }
        },
        "summary": "Import CSV file"
      }
    },
    "/import/docx": {
      "post": {
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportOkResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportErrorResponse"
                }
              }
            },
            "description": "Invalid"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportErrorResponse"
                }
              }
            },
            "description": "Rate limited"
          }
        },
        "summary": "Import DOCX table"
      }
    },
    "/import/menu": {
      "post": {
        "parameters": [
          {
            "description": "If true (1) perform validation + diff only (no persistence)",
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "minimal": {
                  "summary": "Minimal valid payload",
                  "value": {
                    "items": [
                      {
                        "name": "Spaghetti Bolognese"
                      }
                    ]
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ImportMenuRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "dryRun": {
                    "value": {
                      "dry_run": true,
                      "meta": {
                        "count": 1,
                        "dry_run": true,
                        "format": "menu"
                      },
                      "ok": true,
                      "rows": [
                        {
                          "description": "Tomato",
                          "priority": 1,
                          "title": "Soup"
                        }
                      ]
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ImportOkResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportErrorResponse"
                }
              }
            },
            "description": "Invalid"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportErrorResponse"
                }
              }
            },
            "description": "Rate limited"
          }
        },
        "summary": "Import weekly menu (dry-run supported)"
      }
    },
    "/import/xlsx": {
      "post": {
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportOkResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportErrorResponse"
                }
              }
            },
            "description": "Invalid"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportErrorResponse"
                }
              }
            },
            "description": "Rate limited"
          }
        },
        "summary": "Import XLSX spreadsheet"
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}